chore: 清理死代码、恒等判定与静默 catch
移除未使用的 menu_list_style / GridView / ListView、DeviceInfo.ssidDash 与 MediaArgs.allFiles（查看器改用 state.viewer 后已无人传）。
LogFormat 的 c.toInt().toChar()、CameraHttp 的 body?.length、QrScanScreen 的 plane.buffer 空判定在编译期就恒成立；FileSaver 两处 catch 只返回 null，补上原因日志。

+++ composeApp/src/commonMain/kotlin/com/rovecamlink/app/core/log/LogFormat.kt
@@ -292,7 +292,7 @@ object LogFormat {
         }
         val ascii = bytes.take(n).joinToString("") { b ->
             val c = b.toInt() and 0xFF
-            if (c in 0x20..0x7E) c.toInt().toChar().toString() else "."
+            if (c in 0x20..0x7E) c.toChar().toString() else "."
         }
         return "$hex| $ascii" + if (bytes.size > n) " ...(+${bytes.size - n} of ${bytes.size}B)" else ""
     }
